requirements.txt
# python==3.8.5
fastapi>=0.63.0
uvicorn>=0.13.4
pandas>=1.2.4
tqdm
pymysql
requests
fake_useragent
pyyaml